requirements.txt
# Core dependencies
fastapi~=0.104.1
uvicorn[standard]~=0.24.0
sqlmodel~=0.0.14
pydantic~=2.5.0
python-jose[cryptography]~=3.3.0
python-multipart~=0.0.9
pydantic-settings~=2.2.0

paho-mqtt~=2.1.0
alembic~=1.13.1
psycopg2-binary~=2.9.9

# Logging and monitoring
structlog~=23.2.0
prometheus-fastapi-instrumentator~=6.1.0

# Security and certificates
cryptography~=41.0.7

# Development dependencies
pytest~=7.4.3
pytest-asyncio~=0.21.1
httpx~=0.25.2
black~=23.11.0
ruff~=0.1.6
pre-commit~=3.5.0
